Move config class to its own dir
